# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NIAT SPI Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recovery/instructor/sessions:
    get:
      operationId: listInstructorRecoverySessions
      tags: [recovery]
      summary: List the authenticated instructor's assigned recovery sessions
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Assigned recovery sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InstructorRecoverySession"
        "403":
          description: Instructor access required
  /recovery/instructor/curriculum:
    get:
      operationId: listInstructorRecoveryCurriculum
      tags: [recovery]
      summary: List curriculum for the authenticated instructor's scoped subjects
      responses:
        "200":
          description: Scoped recovery curriculum
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InstructorRecoveryCurriculum"
        "403":
          description: Instructor access required
  /recovery/sessions/{id}/report:
    post:
      operationId: reportRecoverySession
      tags: [recovery]
      summary: Submit the authenticated instructor's recovery session report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoverySessionReportInput"
      responses:
        "200":
          description: Submitted recovery report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoverySessionReportResult"
        "403":
          description: Session is not assigned to the caller
        "409":
          description: Session is no longer reportable
...
components:
  schemas:
...
    InstructorRecoveryTopic:
      type: object
      required: [id, sequenceNo, title, order]
      properties:
        id:
          type: string
          format: uuid
        sequenceNo:
          type: integer
        title:
          type: string
        order:
          type: integer
    InstructorRecoverySession:
      type: object
      required: [id, campus, subject, section, scheduledDate, startTime, endTime, studentsExpected, topics]
      properties:
        id:
          type: string
          format: uuid
        campus:
          type: string
        subject:
          type: string
        section:
          type: ["string", "null"]
        scheduledDate:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        studentsExpected:
          type: ["integer", "null"]
        topics:
          type: array
          items:
            $ref: "#/components/schemas/InstructorRecoveryTopic"
    InstructorRecoveryCurriculumTopic:
      type: object
      required: [id, sequenceNo, weekNo, title]
      properties:
        id:
          type: string
          format: uuid
        sequenceNo:
          type: integer
        weekNo:
          type: ["integer", "null"]
        title:
          type: string
    InstructorRecoveryCurriculum:
      type: object
      required: [campus, subject, topics]
      properties:
        campus:
          type: string
        subject:
          type: string
        topics:
          type: array
          items:
            $ref: "#/components/schemas/InstructorRecoveryCurriculumTopic"
    RecoverySessionReportInput:
      type: object
      required: [coveredTopicIds]
      properties:
        coveredTopicIds:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
        studentsAttended:
          type: integer
          minimum: 0
    RecoverySessionReportResult:
      type: object
      required: [id, status, reportedAt]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [conducted, partial, no_show]
        reportedAt:
          type: string
          format: date-time
...