# lib/api-zod/src/generated/types/instructorRecoveryCurriculum.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NIAT SPI Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { InstructorRecoveryCurriculumTopic } from './instructorRecoveryCurriculumTopic';

export interface InstructorRecoveryCurriculum {
  campus: string;
  subject: string;
  topics: InstructorRecoveryCurriculumTopic[];
}