# lib/api-zod/src/generated/types/instructorRecoveryCurriculumTopic.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NIAT SPI Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface InstructorRecoveryCurriculumTopic {
  id: string;
  sequenceNo: number;
  /** @nullable */
  weekNo: number | null;
  title: string;
}