# lib/api-zod/src/generated/types/instructorRecoverySession.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NIAT SPI Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { InstructorRecoveryTopic } from './instructorRecoveryTopic';

export interface InstructorRecoverySession {
  id: string;
  campus: string;
  subject: string;
  /** @nullable */
  section: string | null;
  scheduledDate: Date;
  startTime: string;
  endTime: string;
  /** @nullable */
  studentsExpected: number | null;
  topics: InstructorRecoveryTopic[];
}