# lib/api-zod/src/generated/types/recoverySessionReportInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NIAT SPI Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RecoverySessionReportInput {
  coveredTopicIds: string[];
  /** @minimum 0 */
  studentsAttended?: number;
}