# lib/api-zod/src/generated/types/recoverySessionReportResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NIAT SPI Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RecoverySessionReportResultStatus } from './recoverySessionReportResultStatus';

export interface RecoverySessionReportResult {
  id: string;
  status: RecoverySessionReportResultStatus;
  reportedAt: Date;
}